+++ config.php
@@ -25,16 +25,29 @@
 defined('MOODLE_INTERNAL') || die();
 
 $THEME->name = 'boost_union';
-$THEME->parents = ['boost'];
 $THEME->sheets = [];
 $THEME->editor_sheets = [];
-$THEME->rendererfactory = 'theme_overridden_renderer_factory';
-$THEME->extrascsscallback = 'theme_boost_union_get_extra_scss';
-$THEME->prescsscallback = 'theme_boost_union_get_pre_scss';
-$THEME->precompiledcsscallback = 'theme_boost_union_get_precompiled_css';
+$THEME->editor_scss = ['editor'];
+$THEME->usefallback = true;
 $THEME->scss = function($theme) {
     return theme_boost_union_get_main_scss_content($theme);
 };
+
+// The $THEME->layouts setting is not duplicated here as they are properly inherited from theme_boost.
+
+$THEME->parents = ['boost'];
+$THEME->enable_dock = false;
+$THEME->extrascsscallback = 'theme_boost_union_get_extra_scss';
+$THEME->prescsscallback = 'theme_boost_union_get_pre_scss';
+$THEME->precompiledcsscallback = 'theme_boost_union_get_precompiled_css';
 $THEME->yuicssmodules = array();
-$THEME->usefallback = true;
+$THEME->rendererfactory = 'theme_overridden_renderer_factory';
+$THEME->requiredblocks = '';
+$THEME->addblockposition = BLOCK_ADDBLOCK_POSITION_FLATNAV;
+$THEME->iconsystem = \core\output\icon_system::FONTAWESOME;
 $THEME->haseditswitch = true;
+$THEME->usescourseindex = true;
+// By default, all boost theme do not need their titles displayed.
+$THEME->activityheaderconfig = [
+    'notitle' => true
+];