+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-theme_boost_union
 Changes
 -------
 
+### Unreleased
+
+* 2022-09-28 - Bugfix: Place stylelint-disable-line declaration-no-important correctly, solves #91.
+
 ### v4.0-r2
 
 * 2022-09-27 - Improvement: Align the fallback CSS file with theme_boost.